regular sync

+++ final_project/dbt_energy/models/staging/weather/stg_weather__forecast.sql
@@ -14,14 +14,14 @@ renamed as (
         cast(`date` as timestamp) as `weather_timestamp`, -- Renaming to avoid reserved keyword issues
         -- 2. Renaming a field (using standard snake_case without backticks for safety)
         
-        cast(temperature_2m as float64) as `temperature_2m`,
-        cast(wind_speed_80m as float64) as `wind_speed_80m`,
-        cast(wind_direction_80m as integer) as `wind_direction_80m`,
-        cast(snowfall as float64) as `snowfall`,
-        cast(precipitation as float64) as `precipitation`,
-        cast(shortwave_radiation as float64) as `solar_radiation`,
-        cast(direct_radiation as float64) as `direct_radiation`,
-        cast(diffuse_radiation as float64) as `diffuse_radiation`        
+        cast(temperature_2m as NUMERIC) as `temperature_2m`,
+        cast(wind_speed_80m as NUMERIC) as `wind_speed_80m`,
+        cast(wind_direction_80m as INT64) as `wind_direction_80m`,
+        cast(snowfall as NUMERIC) as `snowfall`,
+        cast(precipitation as NUMERIC) as `precipitation`,
+        cast(shortwave_radiation as NUMERIC) as `solar_radiation`,
+        cast(direct_radiation as NUMERIC) as `direct_radiation`,
+        cast(diffuse_radiation as NUMERIC) as `diffuse_radiation`        
 
     from source
 )
